requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
alembic==1.16.2
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   groq
    #   httpx
    #   starlette
asyncpg==0.30.0
    # via -r requirements.in
bcrypt==4.3.0
    # via -r requirements.in
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
distro==1.9.0
    # via groq
ecdsa==0.19.1
    # via python-jose
fastapi==0.115.14
    # via -r requirements.in
greenlet==3.2.3
    # via -r requirements.in
groq==0.29.0
    # via langchain-groq
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   groq
    #   langsmith
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.26
    # via -r requirements.in
langchain-core==0.3.66
    # via
    #   langchain
    #   langchain-groq
    #   langchain-text-splitters
langchain-groq==0.3.4
    # via -r requirements.in
langchain-text-splitters==0.3.8
    # via langchain
langsmith==0.4.4
    # via
    #   langchain
    #   langchain-core
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
orjson==3.10.18
    # via langsmith
packaging==24.2
    # via
    #   langchain-core
    #   langsmith
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pydantic==2.11.7
    # via
    #   fastapi
    #   groq
    #   langchain
    #   langchain-core
    #   langsmith
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via -r requirements.in
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
python-jose==3.5.0
    # via -r requirements.in
python-slugify==8.0.4
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-core
requests==2.32.4
    # via
    #   -r requirements.in
    #   langchain
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via
    #   anyio
    #   groq
sqlalchemy==2.0.41
    # via
    #   -r requirements.in
    #   alembic
    #   langchain
starlette==0.46.2
    # via fastapi
tenacity==9.1.2
    # via langchain-core
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.14.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   groq
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
uvicorn==0.34.3
    # via -r requirements.in
zstandard==0.23.0
    # via langsmith
